+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.1.0 (Jun 16, 2021)
+
+* minified
+
+## 0.1.0 (Jun 10, 2021)
+
+* disabled debug
+* removed hardcoded usertoken
+* added saving usertoken via cookie, to prevent user needing to enter it every time on the same server
+
 ## 0.1.0 (Jun 10, 2021)
 
 * Initial Release